chore: Rewrite schema structure
